Allowing to disconnect accounts and the bot account

+++ server/command.go
@@ -68,8 +68,14 @@ func getAutocompleteData() *model.AutocompleteData {
 	connect := model.NewAutocompleteData("connect", "", "Connect your Mattermost account to your MS Teams account")
 	cmd.AddCommand(connect)
 
+	disconnect := model.NewAutocompleteData("disconnect", "", "Disconnect your Mattermost account from your MS Teams account")
+	cmd.AddCommand(disconnect)
+
 	connectBot := model.NewAutocompleteData("connect-bot", "", "Connect the bot account (only system admins can do this)")
 	cmd.AddCommand(connectBot)
+
+	disconnectBot := model.NewAutocompleteData("disconnect-bot", "", "Disconnect the bot account (only system admins can do this)")
+	cmd.AddCommand(disconnectBot)
 
 	return cmd
 }
@@ -108,6 +114,14 @@ func (p *Plugin) ExecuteCommand(c *plugin.Context, args *model.CommandArgs) (*mo
 
 	if action == "connect-bot" {
 		return p.executeConnectBotCommand(c, args)
+	}
+
+	if action == "disconnect" {
+		return p.executeDisconnectCommand(c, args)
+	}
+
+	if action == "disconnect-bot" {
+		return p.executeDisconnectBotCommand(c, args)
 	}
 
 	return cmdError(args.ChannelId, "Unknown command. Valid options: link, unlink and show.")
@@ -300,6 +314,45 @@ func (p *Plugin) executeConnectBotCommand(c *plugin.Context, args *model.Command
 	message = <-messageChan
 	p.sendBotEphemeralPost(args.UserId, args.ChannelId, message)
 	close(messageChan)
+	return &model.CommandResponse{}, nil
+}
+
+func (p *Plugin) executeDisconnectCommand(c *plugin.Context, args *model.CommandArgs) (*model.CommandResponse, *model.AppError) {
+	teamsUserID, err := p.store.MattermostToTeamsUserID(args.UserId)
+	if err != nil {
+		p.sendBotEphemeralPost(args.UserId, args.ChannelId, "Error: the account is not connected")
+		return &model.CommandResponse{}, nil
+	}
+
+	err = p.store.SetUserInfo(p.userID, teamsUserID, nil)
+	if err != nil {
+		p.sendBotEphemeralPost(args.UserId, args.ChannelId, fmt.Sprintf("Error: unable to disconnect your account, %s", err.Error()))
+		return &model.CommandResponse{}, nil
+	}
+
+	p.sendBotEphemeralPost(args.UserId, args.ChannelId, "Your account has been disconnected")
+
+	return &model.CommandResponse{}, nil
+}
+
+func (p *Plugin) executeDisconnectBotCommand(c *plugin.Context, args *model.CommandArgs) (*model.CommandResponse, *model.AppError) {
+	if !p.API.HasPermissionTo(args.UserId, model.PermissionManageSystem) {
+		return cmdError(args.ChannelId, "Unable to connect the bot account, only system admins can connect the bot account.")
+	}
+
+	botTeamsUserID, err := p.store.MattermostToTeamsUserID(p.userID)
+	if err != nil {
+		p.sendBotEphemeralPost(args.UserId, args.ChannelId, "Error: unable to find the connected bot account")
+		return &model.CommandResponse{}, nil
+	}
+	err = p.store.SetUserInfo(p.userID, botTeamsUserID, nil)
+	if err != nil {
+		p.sendBotEphemeralPost(args.UserId, args.ChannelId, fmt.Sprintf("Error: unable to disconnect the bot account, %s", err.Error()))
+		return &model.CommandResponse{}, nil
+	}
+
+	p.sendBotEphemeralPost(args.UserId, args.ChannelId, "The bot account has been disconnected")
+
 	return &model.CommandResponse{}, nil
 }
